Tree: Move mark splitting utilities to more appropriate file (#15297)
## Description

Moved splitMarkOnInput and splitMarkOnOutput from a file related for
cross-field move handling to a general sequence change handling utility
file.
