+++ experimental/dds/tree2/src/feature-libraries/sequence-field/index.ts
@@ -73,5 +73,4 @@ export {
 	MoveEffectTable,
 	MoveEffect,
 	PairedMarkUpdate,
-	splitMarkOnOutput,
 } from "./moveEffectTable";
