+++ experimental/dds/tree2/src/feature-libraries/sequence-field/utils.ts
@@ -58,7 +58,6 @@ import {
 	makeMergeable,
 	MoveEffectTable,
 	MoveMark,
-	splitMarkOnOutput,
 	splitMove,
 } from "./moveEffectTable";
 
@@ -929,4 +928,208 @@ export function newCrossFieldTable<T = unknown>(): CrossFieldTable<T> {
  */
 export function newMoveEffectTable<T>(): MoveEffectTable<T> {
 	return newCrossFieldTable();
+}
+
+/**
+ * Splits the `mark` into two marks such that the first returned mark has input length `length`.
+ * @param mark - The mark to split.
+ * @param revision - The revision of the changeset the mark is part of.
+ * @param length - The desired length for the first of the two returned marks.
+ * @param genId - An ID allocator
+ * @param moveEffects - The table in which to record splitting of move marks
+ * @param recordMoveEffect - Whether when splitting a move an entry should be added to `moveEffects` indicating that the mark should be split (in case we process this mark again).
+ * An entry is always added to `moveEffects` indicating that the opposite end of the move should be split.
+ * @returns A pair of marks equivalent to the original `mark`
+ * such that the first returned mark has input length `length`.
+ */
+export function splitMarkOnInput<T, TMark extends InputSpanningMark<T>>(
+	mark: TMark,
+	revision: RevisionTag | undefined,
+	length: number,
+	genId: IdAllocator,
+	moveEffects: MoveEffectTable<T>,
+	recordMoveEffect: boolean = false,
+): [TMark, TMark] {
+	const markLength = getInputLength(mark);
+	const remainder = markLength - length;
+	if (length < 1 || remainder < 1) {
+		fail("Unable to split mark due to lengths");
+	}
+	if (isSkipMark(mark)) {
+		return [length, remainder] as [TMark, TMark];
+	}
+	const markObj = mark as Exclude<TMark, Skip>;
+	const type = mark.type;
+	switch (type) {
+		case "Modify":
+			fail("Unable to split Modify mark of length 1");
+		case "ReturnTo": {
+			const newId = genId();
+			splitMove(
+				moveEffects,
+				CrossFieldTarget.Source,
+				mark.revision ?? revision,
+				mark.id,
+				newId,
+				length,
+				remainder,
+			);
+			if (recordMoveEffect) {
+				splitMove(
+					moveEffects,
+					CrossFieldTarget.Destination,
+					mark.revision ?? revision,
+					mark.id,
+					newId,
+					length,
+					remainder,
+				);
+			}
+			return [
+				{ ...markObj, count: length },
+				{ ...markObj, id: newId, count: remainder, detachIndex: mark.detachIndex + length },
+			] as [TMark, TMark];
+		}
+		case "Revive":
+			return [
+				{ ...markObj, content: mark.content.slice(0, length), count: length },
+				{
+					...markObj,
+					content: mark.content.slice(length),
+					count: remainder,
+					detachIndex: mark.detachIndex + length,
+				},
+			] as [TMark, TMark];
+		case "Delete":
+			return [
+				{ ...markObj, count: length },
+				{ ...markObj, count: remainder },
+			] as [TMark, TMark];
+		case "MoveOut":
+		case "ReturnFrom": {
+			// TODO: Handle detach index for ReturnFrom
+			const newId = genId();
+			splitMove(
+				moveEffects,
+				CrossFieldTarget.Destination,
+				mark.revision ?? revision,
+				mark.id,
+				newId,
+				length,
+				remainder,
+			);
+			if (recordMoveEffect) {
+				splitMove(
+					moveEffects,
+					CrossFieldTarget.Source,
+					mark.revision ?? revision,
+					mark.id,
+					newId,
+					length,
+					remainder,
+				);
+			}
+			const mark1 = { ...markObj, count: length };
+			const mark2 = { ...markObj, id: newId, count: remainder };
+			if (mark.type === "ReturnFrom" && mark.detachIndex !== undefined) {
+				(mark2 as unknown as ReturnFrom<unknown>).detachIndex = mark.detachIndex + length;
+			}
+			return [mark1, mark2];
+		}
+		default:
+			unreachableCase(type);
+	}
+}
+
+/**
+ * Splits the `mark` into two marks such that the first returned mark has output length `length`.
+ * @param mark - The mark to split.
+ * @param revision - The revision of the changeset the mark is part of.
+ * @param length - The desired length for the first of the two returned marks.
+ * @param genId - An ID allocator
+ * @param moveEffects - The table in which to record splitting of move marks
+ * @param recordMoveEffect - Whether when splitting a move an entry should be added to `moveEffects` indicating that the mark should be split (in case we process this mark again).
+ * An entry is always added to `moveEffects` indicating that the opposite end of the move should be split.
+ * @returns A pair of marks equivalent to the original `mark`
+ * such that the first returned mark has output length `length`.
+ */
+export function splitMarkOnOutput<T, TMark extends OutputSpanningMark<T>>(
+	mark: TMark,
+	revision: RevisionTag | undefined,
+	length: number,
+	genId: IdAllocator,
+	moveEffects: MoveEffectTable<T>,
+	recordMoveEffect: boolean = false,
+	ignorePairing: boolean = false,
+): [TMark, TMark] {
+	const markLength = getOutputLength(mark, ignorePairing);
+	const remainder = markLength - length;
+	if (length < 1 || remainder < 1) {
+		fail("Unable to split mark due to lengths");
+	}
+	if (isSkipMark(mark)) {
+		return [length, remainder] as [TMark, TMark];
+	}
+	const markObj = mark as Exclude<TMark, Skip>;
+	const type = markObj.type;
+	switch (type) {
+		case "Modify":
+			fail(`Unable to split ${type} mark of length 1`);
+		case "Insert":
+			return [
+				{ ...markObj, content: markObj.content.slice(0, length) },
+				{
+					...markObj,
+					content: markObj.content.slice(length),
+					id: (markObj.id as number) + length,
+				},
+			];
+		case "MoveIn":
+		case "ReturnTo": {
+			const newId = genId();
+			splitMove(
+				moveEffects,
+				CrossFieldTarget.Source,
+				markObj.revision ?? revision,
+				markObj.id,
+				newId,
+				length,
+				remainder,
+			);
+			if (recordMoveEffect) {
+				splitMove(
+					moveEffects,
+					CrossFieldTarget.Destination,
+					markObj.revision ?? revision,
+					markObj.id,
+					newId,
+					length,
+					remainder,
+				);
+			}
+			return [
+				{ ...markObj, count: length },
+				type === "MoveIn"
+					? { ...markObj, id: newId, count: remainder }
+					: {
+							...markObj,
+							id: newId,
+							count: remainder,
+							detachIndex: markObj.detachIndex + length,
+					  },
+			];
+		}
+		case "Revive":
+			return [
+				{ ...markObj, content: markObj.content.slice(0, length), count: length },
+				{
+					...markObj,
+					content: markObj.content.slice(length),
+					count: remainder,
+					detachIndex: markObj.detachIndex + length,
+				},
+			];
+		default:
+			unreachableCase(type);
+	}
 }
